Added FLV support in FLU file requester. Added UI color support in preferences (albeit it refreshes only when saved).

+++ mrViewer/src/core/mrvServer.h
@@ -118,7 +118,7 @@ struct ServerData
 {
      std::string host;
      std::string group;
-     unsigned    port;
+     unsigned short port;
      mrv::ViewerUI* ui;
 };
